workspace/implementation/requirements.txt
playwright>=1.40.0
aiohttp>=3.9.0
click>=8.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
